requirements.txt
opencv-python
numpy
pyautogui

Pillow
tk
